requirements.txt
pulp
pandas
streamlit
plotly
oracledb
PyYAML
fastapi
uvicorn
